Fix use of uninitialized zowelog.
zowelog() cannot be used without a proper environment. The main function
of ZSS was changed to use zowelog instead of printf which led to use of
zowelog() before its environment was initialized, which in turn led to a
segfault.

This commit moves all the zowelog invocations to a place where the
logging environment is ready. Additionally, clean up logic has been
introduced to ensure that we free the STC base resources before leaving
main().

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>