AUX communication PC disablement
With the introduction of the REUSASID=NO option for AUX, it became
somewhat dangerous to run it with the communication PC as it would leak
ASIDs. This commit adds the ability to disable the communication PC and
introduces an alternative communication path for the termination signal.

Major changes:
* Disablement of the communication PC
* Termination signal via a cross-memory POST

Minor changes:
* Defines for AUX parameters
* Communication area versioning

Fixes: #250

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>