Pass ZIS version to dynlink and use it for ZIS check
This commit changes the dynamic linkage base plugin
build script to pass the ZIS version to the plugin.
The plugin then uses that version to check whether
the host ZIS is compatible with it. This also fixes,
the ZWES0214E message since it's expected those ZIS
version #define's to be present (they weren't before
this commit hence incorrect values in the message).

Fixes: #586

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>