Change conflicting message IDs
This commit changes the conflicting messages IDs in the range 200-214
to be 700-714. According to zowe-common-c/h/crossmemory.h 700-999 are
for user service which are very unlikely to be implemented since
ZIS is extended via plugins now, therefore we can use that range (that
should be noted in zowe-common-c/h/crossmemory.h).

Fixes: #590

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>