Disable the ZIS dynamic linkage plugin
The dynamic linkage plugin isn't needed by
default, so we shouldn't have it enabled as
it uses limited z/OS resources (LPA).

Also, remove an incorrect comment in the SAMPLIB:
the dynamic linkage plugin does not have to be
first, it's loaded first by ZIS automatically
regardless of its position in the config.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>