Utilize the module registry
To decrease the ZIS LPA footprint, ZIS will use the module registry for
its plug-in modules. When not in dev mode, ZIS will use the module
registry API for adding plug-in modules to the LPA, and if a module is
already registered, ZIS will reuse the existing LPA info. In dev mode,
ZIS will load its own copy of plug-in modules to the LPA, just as it
does now.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>