Make sure shared modules are never deleted
This commit introduces a new flag which shows if a plug-in module is
private and is safe to be deleted when the dev options are used.

The corresponding modreg changes in zowe-common-c, which are picked up in
this commit, will ensure that the LPMEA area from modregRegister never
contains a valid delete token, which means it can't be used to delete
the corresponding module.

Additionally, in dev mode, print a message explicitly saying shared plug-in
modules aren't deleted.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>